# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Raj Hotel Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  customer: "customer",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface SignupRequest {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Room {
  id: number;
  title: string;
  description: string;
  price: number;
  imageUrl: string | null;
  isAvailable: boolean;
  createdAt: string;
}

export interface CreateRoomRequest {
  title: string;
  description: string;
  price: number;
  imageUrl?: string | null;
  isAvailable: boolean;
}

export interface UpdateRoomRequest {
  title?: string;
  description?: string;
  price?: number;
  imageUrl?: string | null;
  isAvailable?: boolean;
}

export interface UploadImageRequest {
  imageBase64: string;
  folder?: string;
}

export interface UploadImageResponse {
  url: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  rejected: "rejected",
} as const;

export interface Booking {
  id: number;
  userId: number;
  roomId: number;
  checkIn: string;
  checkOut: string;
  status: BookingStatus;
  createdAt: string;
}

export type BookingWithDetailsStatus =
  (typeof BookingWithDetailsStatus)[keyof typeof BookingWithDetailsStatus];

export const BookingWithDetailsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  rejected: "rejected",
} as const;

export interface BookingWithDetails {
  id: number;
  userId: number;
  roomId: number;
  checkIn: string;
  checkOut: string;
  status: BookingWithDetailsStatus;
  createdAt: string;
  user: User;
  room: Room;
}

export interface CreateBookingRequest {
  roomId: number;
  checkIn: string;
  checkOut: string;
}

export type UpdateBookingRequestStatus =
  (typeof UpdateBookingRequestStatus)[keyof typeof UpdateBookingRequestStatus];

export const UpdateBookingRequestStatus = {
  pending: "pending",
  confirmed: "confirmed",
  rejected: "rejected",
} as const;

export interface UpdateBookingRequest {
  status: UpdateBookingRequestStatus;
}

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const PaymentStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Payment {
  id: number;
  bookingId: number;
  amount: number;
  screenshotUrl: string | null;
  status: PaymentStatus;
  createdAt: string;
}

export type PaymentWithDetailsStatus =
  (typeof PaymentWithDetailsStatus)[keyof typeof PaymentWithDetailsStatus];

export const PaymentWithDetailsStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface PaymentWithDetails {
  id: number;
  bookingId: number;
  amount: number;
  screenshotUrl: string | null;
  status: PaymentWithDetailsStatus;
  createdAt: string;
  booking: BookingWithDetails;
}

export interface CreatePaymentRequest {
  bookingId: number;
  amount: number;
  screenshotUrl?: string | null;
}

export type UpdatePaymentRequestStatus =
  (typeof UpdatePaymentRequestStatus)[keyof typeof UpdatePaymentRequestStatus];

export const UpdatePaymentRequestStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdatePaymentRequest {
  status: UpdatePaymentRequestStatus;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  delivered: "delivered",
} as const;

export interface Order {
  id: number;
  userId: number;
  itemName: string;
  price: number;
  status: OrderStatus;
  createdAt: string;
}

export interface CreateOrderRequest {
  itemName: string;
  price: number;
}

export type UpdateOrderRequestStatus =
  (typeof UpdateOrderRequestStatus)[keyof typeof UpdateOrderRequestStatus];

export const UpdateOrderRequestStatus = {
  pending: "pending",
  delivered: "delivered",
} as const;

export interface UpdateOrderRequest {
  status: UpdateOrderRequestStatus;
}

export interface MonthlyRevenue {
  month: string;
  revenue: number;
}

export interface BookingTrend {
  month: string;
  count: number;
}

export interface AnalyticsData {
  totalBookings: number;
  totalRevenue: number;
  activeRooms: number;
  pendingPayments: number;
  monthlyRevenue: MonthlyRevenue[];
  bookingTrends: BookingTrend[];
}